# trading-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Trading Service API",
        "title": "Trading Service",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/orders": {
            "get": {
                "description": "Vraća listu naloga, opciono filtriranu po statusu.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Preuzimanje liste naloga",
                "parameters": [
                    {
                        "type": "string",
                        "default": "all",
                        "example": "pending",
                        "description": "Status naloga za filtriranje. Podrazumevano 'all' za sve statuse.",
                        "name": "filter_status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno preuzeta lista naloga",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.OrderResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Greška pri preuzimanju naloga iz baze",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Kreira novi nalog za hartije od vrednosti.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Kreiranje novog naloga",
                "parameters": [
                    {
                        "description": "Podaci neophodni za kreiranje naloga",
                        "name": "orderRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Uspešno kreiran nalog, vraća ID novog naloga",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Neispravan format, neuspela validacija ili greška pri upisu u bazu",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "403": {
                        "description": "Nije dozvoljeno kreirati nalog za drugog korisnika",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "description": "Vraća detalje specifičnog naloga na osnovu njegovog jedinstvenog identifikatora (ID).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Preuzimanje naloga po ID-u",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID naloga koji se preuzima",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno preuzet nalog",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.OrderResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Nevalidan ID naloga",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "404": {
                        "description": "Nalog sa datim ID-jem ne postoji",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/orders/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menja status naloga u 'approved'.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Odobravanje naloga",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID naloga koji se odobrava",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nalog uspešno odobren",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Nevalidan ID ili nalog nije u 'pending' statusu",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "403": {
                        "description": "Nedovoljne privilegije",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "404": {
                        "description": "Nalog sa datim ID-jem ne postoji",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Interna Greška Servera",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.CreateOrderRequest": {
            "type": "object",
            "required": [
                "account_id",
                "contract_size",
                "direction",
                "quantity",
                "security_id",
                "user_id"
            ],
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "aon": {
                    "type": "boolean"
                },
                "contract_size": {
                    "type": "integer"
                },
                "direction": {
                    "type": "string"
                },
                "limit_price_per_unit": {
                    "type": "number"
                },
                "margin": {
                    "type": "boolean"
                },
                "quantity": {
                    "type": "integer"
                },
                "security_id": {
                    "type": "integer"
                },
                "stop_price_per_unit": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "types.OrderResponse": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "after_hours": {
                    "type": "boolean"
                },
                "aon": {
                    "type": "boolean"
                },
                "approved_by": {
                    "description": "Supervizor koji je odobrio order",
                    "type": "integer"
                },
                "contract_size": {
                    "type": "integer"
                },
                "direction": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_done": {
                    "type": "boolean"
                },
                "last_modified": {
                    "type": "integer"
                },
                "limit_price_per_unit": {
                    "type": "number"
                },
                "margin": {
                    "type": "boolean"
                },
                "quantity": {
                    "type": "integer"
                },
                "remaining_parts": {
                    "type": "integer"
                },
                "security_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "stop_price_per_unit": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "types.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Unesite token. Primer: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}